requirements.txt
asgiref==3.7.2
Django==5.0.3
Pillow==10.1.0
sqlparse==0.4.4
typing_extensions==4.8.0
gunicorn==20.1.0
django-tinymce==4.0.0